feat: add user-agent rotation middleware and import it in settings
