+++ deep_research/komkom_scraper/settings.py
@@ -30,19 +30,4 @@ RETRY_TIMES = 3
 LOG_LEVEL = "INFO"
 
 # Custom middleware for user-agent rotation
-import random
-from scrapy import signals
-
-class RandomUserAgentMiddleware:
-    def __init__(self, user_agents):
-        self.user_agents = user_agents
-
-    @classmethod
-    def from_crawler(cls, crawler):
-        return cls(
-            user_agents=crawler.settings.get('USER_AGENTS', [])
-        )
-
-    def process_request(self, request, spider):
-        if self.user_agents:
-            request.headers.setdefault('User-Agent', random.choice(self.user_agents))
+from komkom_scraper.middlewares import RandomUserAgentMiddleware
